crewai-workspace/news-reader-agent/pyproject.toml
[project]
name = "news-reader-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "crewai[tools]>=0.152.0",
    "playwright>=1.58.0",
    "python-dotenv>=1.1.1",
]
